Change Bezier to_svg to include handles and endpoints
