+++ libraries/bezier-rs/src/lib.rs
@@ -3,7 +3,9 @@
 mod bezier;
 mod consts;
 mod subpath;
+mod svg;
 mod utils;
 
 pub use bezier::*;
 pub use subpath::*;
+pub use svg::*;
